Force azure provider registration to false.
Azure provider registration allow terraform provider to automatically
register azure resource providers.
In our context we are running read only so we do not want to try to
enable them during the scan.
Without this, a driftctl scan will try to register resource providers
since it is done in the gRPC configure call.