Disambiguate between drift and error exit codes
`driftctl scan` exits with zero only when there is no detected drift.
Prior to this commit, it exited with 1 when there was a non-panic error,
and also when there was detected drift. This makes error handling in
scripts more awkward than it needs to be.

This commit changes the exit code to 2 whenever there are errors during
a scan, whether that error is from a panic or not, leaving 1 as the exit
code used solely for when drift is detected.

Related: https://github.com/snyk/driftctl/issues/1350