chore: include uncategorised changes in changelog
Ensure all PRs, even ones missing a "kind" label, are present in the
generated changelog. During the manually-initiated release process,
maintainers can edit the changelog to categorise them if they wish.

If we move to an automated, continuous release, it will still be better
to include uncategorised PRs than to omit them altogether.