fix: fix broken tests
Those tests were broken due to removal of SetResolveReadAttributesFunc.
Sometimes in those methods we were casting field from different types to string.
If we loose that case it causes some issues since in the detail fetcher we only take into account
strings attributes. To fix that I added some cast directly in the detail fetcher.
That should not cause any issues to retrieve details if we send additional (useless) fields to the
ReadResource call.