fix: enumlib: remove a scan-specific error message
Since AWS is the default cloud provider for `driftctl scan`, we give
users a hint on how to configure different providers with `--to` if
there is an error authenticating with AWS - since they might not have
even intended to use AWS in the first place.

Now that this functionality is inside the enumeration library, with
consumers other than `driftctl scan`, we remove this hint from the lower
layer and re-add it for the scan command only, to avoid confusing
non-driftctl users.