chore: don't disable acc test KMS keys
When applying acceptance test terraform for AWS KMS key resources, we
often see errors of the form: `Failed to set KMS key ... status to
false: "NotFoundException: Key '...' does not exist"`. Our current
theory is that setting is_enabled=false (as we currently do) is causing
terraform to have to perform 2 API operations: create the key, then
disable it. Due to the "eventualy consistency" of this API we see this
flaking quite often.

We don't actually need to disable the keys, so let's not, in an attempt
to improve the reliability of this test.