chore: stream acceptance test output in CI
We see reasonably frequent acceptance flakes in CircleCI due to out job
being terminated due to 10 minutes passing without output:
https://support.circleci.com/hc/en-us/articles/360045268074-Build-Fails-with-Too-long-with-no-output-exceeded-10m0s-context-deadline-exceeded-

As well as making that test flaky, the SIGKILL (which does not give us a
chance to clean up, even if we handled other signals during tests),
often causes resources to be left in the acc testing cloud accounts,
producing further failures on the next run, and manual cleanup work for
the team.

We could adjust this time limit, but this commit changes the gotestsum
output to be streaming (like `go test -v`) instead. We still output a
junit xml with gotestsum to be able to look at structured failures in
CircleCI.

We've recently introduced some long backoff/retry loops to try to squash
other flakes, and it would be useful to see which of these are taking a
long time in the test output.